+++ Magic/src/main/java/com/elmakers/mine/bukkit/world/populator/builtin/FrequencyPopulator.java
@@ -1,7 +1,5 @@
 package com.elmakers.mine.bukkit.world.populator.builtin;
 
-import java.util.ArrayList;
-import java.util.List;
 import java.util.Random;
 
 import org.bukkit.Location;
@@ -11,29 +9,32 @@ import org.bukkit.generator.WorldInfo;
 
 import com.elmakers.mine.bukkit.world.populator.BaseBlockPopulator;
 
-public class SequencePopulator extends BaseBlockPopulator {
-    private List<BaseBlockPopulator> populators = new ArrayList<>();
+public class FrequencyPopulator extends BaseBlockPopulator {
+    private BaseBlockPopulator populator;
+    private int frequency = 2;
 
-    @Override
     public boolean onLoad(ConfigurationSection config) {
-        populators = BaseBlockPopulator.loadPopulators(world, config);
-        return !populators.isEmpty();
+        frequency = config.getInt("frequency", frequency);
+        String populatorId = config.getString("populator");
+        populator = BaseBlockPopulator.loadPopulator(world, populatorId);
+        return populator != null;
+    }
+
+    protected boolean isValid(int chunkX, int chunkZ) {
+        return chunkX % frequency == 0 && chunkZ % frequency == 0;
     }
 
     @Override
     public void populate(WorldInfo worldInfo, Random random, int chunkX, int chunkZ, LimitedRegion region) {
-        for (BaseBlockPopulator populator : populators) {
+        if (isValid(chunkX, chunkZ)) {
             populator.populate(worldInfo, random, chunkX, chunkZ, region);
         }
     }
 
     @Override
     public String getPortalTargetWorld(Location location) {
-        for (BaseBlockPopulator populator : populators) {
-            String targetWorld = populator.getPortalTargetWorld(location);
-            if (targetWorld != null) {
-                return targetWorld;
-            }
+        if (isValid(location.getChunk().getX(), location.getChunk().getZ())) {
+            return populator.getPortalTargetWorld(location);
         }
         return null;
     }